third_party/blink/web_tests/platform/generic/fast/forms/select/menulist-disabled-selected-option-expected.txt

Check that a select control does not produce a form value if the selected option element is disabled.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS query.indexOf("select=Disabled") is -1
PASS query.indexOf("select=Enabled") is -1
PASS successfullyParsed is true

TEST COMPLETE

Disabled
Enabled
